Modified classes so that DisplayProfile is split up into 2 different views: common and friend
